\name{ChemometricsWithR-package}
\Rdversion{1.1}
\alias{ChemometricsWithR-package}
\alias{ChemometricsWithR}
\docType{package}
\title{Chemometrics with R - Multivariate Data Analysis in the Natural
  Sciences and Life Sciences}
\description{Package containing scripts and functions from the book
  "Chemometrics with R - Multivariate Data Analysis in the Natural
  Sciences and Life Sciences", by Ron Wehrens. All R code in the book is
  available in demos: typing \code{demo(chapter2)} will run the examples
  from the second chapter. A list of errata can be found in the
  \code{inst} directory of the installed package.
}
\author{Ron Wehrens <ron.wehrens@iasma.it>}
\references{
  R. Wehrens. "Chemometrics with R - Multivariate Data Analysis in the
  Natural Sciences and Life Sciences". Springer, Heidelberg, 2011.
}
\keyword{ package }
\seealso{\code{\link[chemometrics:chemometrics-package]{chemometrics}}}
