\name{unsigned.range}
\alias{unsigned.range}
\title{Unsigned range of the data vector including zero.}
\description{
Function returning the range of the data where, if necessary, the range
is extended to include zero. Not meant to be called directly by the user.
}
\usage{
unsigned.range(x)
}
\arguments{
  \item{x}{Numeric vector.}
}
\value{A vector of two numbers.}
\note{
From the R stats package (see biplot.default).
}
\keyword{misc}
