\name{ChemometricsWithR-package}
\alias{ChemometricsWithR-package}
\alias{ChemometricsWithR}
\docType{package}
\title{
\packageTitle{ChemometricsWithR}
}
\description{
\packageDescription{ChemometricsWithR}
}
\details{
  Package accompanying the second edition of the book
  "Chemometrics with R". The package will not be hosted on CRAN
  and therefore no longer has to comply to the size constraints that
  CRAN has imposed. This simplifies matters considerably for readers and
  in particular removes the need to install a separate data package.

  The scripts in the \code{demo} directory can be used to reproduce the
  results and plots of the individual chapters. For Chapter 3, for
  example, simply run \code{demo("chapter3.R")}.
}
\author{
\packageAuthor{ChemometricsWithR}

Maintainer: \packageMaintainer{ChemometricsWithR}
}
\references{
 R. Wehrens. "Chemometrics with R - Multivariate Data Analysis in the
 Natural Sciences and Life Sciences", 2nd edition, Springer, Heidelberg, 2019.

}
\keyword{ package }
