\name{ChemometricsWithR-package}
\alias{ChemometricsWithR-package}
\alias{ChemometricsWithR}
\docType{package}
\title{
\packageTitle{ChemometricsWithR}
}
\description{
\packageDescription{ChemometricsWithR}
}
\details{
The accompanying \pkg{ChemometricsWithRData} package (approx. 60
MB) contains a couple of data sets used in the book that are too large
for CRAN according to current size limits. For details on how to install
the data package, use \code{?installChemometricsWithRData}.
}
\author{
\packageAuthor{ChemometricsWithR}

Maintainer: \packageMaintainer{ChemometricsWithR}
}
\references{
 R. Wehrens. "Chemometrics with R - Multivariate Data Analysis in the
 Natural Sciences and Life Sciences". Springer, Heidelberg, 2011.

}
\keyword{ package }
